better cmake - since the previous experimen was that successfull, let's see if we can repeat the pattern
