+++ OpenHD/ohd_common/CMakeLists.txt
@@ -44,6 +44,8 @@ set(sources
     "src/openhd_util.cpp"
     "src/openhd_util_filesystem.cpp"
     "src/openhd_settings_persistent.cpp"
+    "src/openhd_profile.cpp"
+    "src/openhd_platform.cpp"
 
     "inc/openhd_settings_imp.hpp"
     "inc/include_json.hpp"
@@ -53,11 +55,8 @@ set(sources
     "inc/openhd_led_codes.hpp"
     "inc/openhd_led_pi.hpp"
     "inc/openhd_link_statistics.hpp"
-    "inc/openhd_platform.hpp"
-    "inc/openhd_platform_discover.hpp"
-    "inc/openhd_platform_json.hpp"
-    "inc/openhd_profile.hpp"
-    "inc/openhd_profile_json.hpp"
+    "inc/openhd_platform.h"
+    "inc/openhd_profile.h"
     "inc/openhd_settings_directories.hpp"
     "inc/openhd_settings_persistent.h"
     "inc/openhd_util.h"
