+++ OpenHD/ohd_interface/inc/ohd_interface.h
@@ -13,8 +13,8 @@
 #include "openhd_action_handler.hpp"
 #include "openhd_external_device.hpp"
 #include "openhd_led_codes.hpp"
-#include "openhd_platform.hpp"
-#include "openhd_profile.hpp"
+#include "openhd_platform.h"
+#include "openhd_profile.h"
 #include "openhd_settings_imp.hpp"
 #include "openhd_spdlog.hpp"
 #include "usb_tether_listener.h"
