+++ OpenHD/ohd_interface/inc/wb_link.h
@@ -13,8 +13,8 @@
 #include "openhd_action_handler.hpp"
 #include "openhd_link.hpp"
 #include "openhd_link_statistics.hpp"
-#include "openhd_platform.hpp"
-#include "openhd_profile.hpp"
+#include "openhd_platform.h"
+#include "openhd_profile.h"
 #include "openhd_settings_imp.hpp"
 #include "openhd_spdlog.hpp"
 #include "wb_link_settings.hpp"
