+++ OpenHD/ohd_interface/inc/wb_link_settings.hpp
@@ -5,7 +5,7 @@
 #ifndef OPENHD_OPENHD_OHD_INTERFACE_INC_WB_LINK_SETTINGS_HPP_
 #define OPENHD_OPENHD_OHD_INTERFACE_INC_WB_LINK_SETTINGS_HPP_
 
-#include "openhd_platform.hpp"
+#include "openhd_platform.h"
 #include "openhd_settings_persistent.h"
 #include "wifi_card.hpp"
