+++ OpenHD/ohd_interface/test/test_led.cpp
@@ -6,7 +6,6 @@
 #include <thread>
 
 #include "openhd_led_codes.hpp"
-#include "openhd_platform_discover.hpp"
 #include "openhd_util.h"
 
 int main(int argc, char *argv[]) {
