+++ OpenHD/ohd_telemetry/tests/test_ground_locally.cpp
@@ -6,9 +6,8 @@
 #include <thread>
 
 #include "../src/OHDTelemetry.h"
-#include "openhd_platform.hpp"
-#include "openhd_platform_discover.hpp"
-#include "openhd_profile.hpp"
+#include "openhd_platform.h"
+#include "openhd_profile.h"
 #include "openhd_settings_imp.hpp"
 
 int main() {
