+++ OpenHD/ohd_video/inc/openhd-rpi-os-configure-vendor-cam.hpp
@@ -9,7 +9,7 @@
 #include <fstream>
 #include <iostream>
 
-#include "openhd_platform.hpp"
+#include "openhd_platform.h"
 #include "openhd_spdlog.hpp"
 #include "openhd_util.h"
 #include "openhd_util_filesystem.h"
